move cli to utils folder
